don't override HTTPRequestSpec
The curl code generator modified the pointer arg, leading to issues
when the generator was substituted for say go afterwards.

Rather than the expected `http.NewRequest("HEAD", url, payload)`
it rendered it as `http.NewRequest("--head", url, payload)`.

Instead, do the sanitization of the request in the applyVariables func
and do the curl specific head transform on a copy.